backend/pyproject.toml
[project]
name = "timetable-backend"
version = "1.0.0"
description = "Timetable Generator SaaS Backend"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.14",
    "djangorestframework-simplejwt>=5.3",
    "django-cors-headers>=4.3",
    "django-filter>=23.5",
    "psycopg2-binary>=2.9",
    "celery>=5.3",
    "redis>=5.0",
    "openpyxl>=3.1",
    "pandas>=2.1",
    "weasyprint>=60.0",
    "python-dotenv>=1.0",
    "drf-spectacular>=0.27",
    "gunicorn>=21.2",
    "whitenoise>=6.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-django>=4.7",
    "pytest-cov>=4.1",
    "black>=23.12",
    "isort>=5.13",
    "ruff>=0.1",
    "factory-boy>=3.3",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.ruff]
line-length = 88
target-version = "py311"
